% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alabamatract20.R
\docType{data}
\name{alabamatract20}
\alias{alabamatract20}
\title{Alabama}
\format{
A data frame with 1437 rows and 398 variables:
}
\source{
/url{https://www2.census.gov/programs-surveys/decennial/2020/data/01-Redistricting_File--PL_94-171/}
}
\usage{
data(alabamatract20)
}
\description{
Tracts for the state of Alabama
}
\keyword{datasets}
