% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alaskatract20.R
\docType{data}
\name{alaskatract20}
\alias{alaskatract20}
\title{Alaska}
\format{
A data frame with 177 rows and 398 variables:
}
\source{
/url{https://www2.census.gov/programs-surveys/decennial/2020/data/01-Redistricting_File--PL_94-171/}
}
\usage{
data(alaskatract20)
}
\description{
Tracts for the state of Alaska
}
\keyword{datasets}
