% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arizonatract20.R
\docType{data}
\name{arizonatract20}
\alias{arizonatract20}
\title{Arizona}
\format{
A data frame with 1765 rows and 398 variables:
}
\source{
/url{https://www2.census.gov/programs-surveys/decennial/2020/data/01-Redistricting_File--PL_94-171/}
}
\usage{
data(arizonatract20)
}
\description{
Tracts for the state of Arizona
}
\keyword{datasets}
