% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arkansastract20.R
\docType{data}
\name{arkansastract20}
\alias{arkansastract20}
\title{Arkansas}
\format{
A data frame with 823 rows and 398 variables:
}
\source{
/url{https://www2.census.gov/programs-surveys/decennial/2020/data/01-Redistricting_File--PL_94-171/}
}
\usage{
data(arkansastract20)
}
\description{
Tracts for the state of Arkansas
}
\keyword{datasets}
