% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coloradotract20.R
\docType{data}
\name{coloradotract20}
\alias{coloradotract20}
\title{Colorado}
\format{
A data frame with 1447 rows and 398 variables:
}
\source{
/url{https://www2.census.gov/programs-surveys/decennial/2020/data/01-Redistricting_File--PL_94-171/}
}
\usage{
data(coloradotract20)
}
\description{
Tracts for the state of Colorado
}
\keyword{datasets}
