% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connecticuttract20.R
\docType{data}
\name{connecticuttract20}
\alias{connecticuttract20}
\title{Connecticut}
\format{
A data frame with 883 rows and 398 variables:
}
\source{
/url{https://www2.census.gov/programs-surveys/decennial/2020/data/01-Redistricting_File--PL_94-171/}
}
\usage{
data(connecticuttract20)
}
\description{
Tracts for the state of Connecticut
}
\keyword{datasets}
