% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dctract20.R
\docType{data}
\name{dctract20}
\alias{dctract20}
\title{District of columbia}
\format{
A data frame with 206 rows and 398 variables:
}
\source{
/url{https://www2.census.gov/programs-surveys/decennial/2020/data/01-Redistricting_File--PL_94-171/}
}
\usage{
data(dctract20)
}
\description{
Tracts for District of columbia
}
\keyword{datasets}
