% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delawaretract20.R
\docType{data}
\name{delawaretract20}
\alias{delawaretract20}
\title{Delaware}
\format{
A data frame with 262 rows and 398 variables:
}
\source{
/url{https://www2.census.gov/programs-surveys/decennial/2020/data/01-Redistricting_File--PL_94-171/}
}
\usage{
data(delawaretract20)
}
\description{
Tracts for the state of Delaware
}
\keyword{datasets}
