% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/floridatract20.R
\docType{data}
\name{floridatract20}
\alias{floridatract20}
\title{Florida}
\format{
A data frame with 5160 rows and 398 variables:
}
\source{
/url{https://www2.census.gov/programs-surveys/decennial/2020/data/01-Redistricting_File--PL_94-171/}
}
\usage{
data(floridatract20)
}
\description{
Tracts for the state of Florida
}
\keyword{datasets}
