% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/georgiatract20.R
\docType{data}
\name{georgiatract20}
\alias{georgiatract20}
\title{Georgia}
\format{
A data frame with 2796 rows and 398 variables:
}
\source{
/url{https://www2.census.gov/programs-surveys/decennial/2020/data/01-Redistricting_File--PL_94-171/}
}
\usage{
data(georgiatract20)
}
\description{
Tracts for the state of Georgia
}
\keyword{datasets}
