% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hawaiitract20.R
\docType{data}
\name{hawaiitract20}
\alias{hawaiitract20}
\title{Hawaii}
\format{
A data frame with 461 rows and 398 variables:
}
\source{
/url{https://www2.census.gov/programs-surveys/decennial/2020/data/01-Redistricting_File--PL_94-171/}
}
\usage{
data(hawaiitract20)
}
\description{
Tracts for the state of Hawaii
}
\keyword{datasets}
