% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iowatract20.R
\docType{data}
\name{iowatract20}
\alias{iowatract20}
\title{Iowa}
\format{
A data frame with 896 rows and 398 variables:
}
\source{
/url{https://www2.census.gov/programs-surveys/decennial/2020/data/01-Redistricting_File--PL_94-171/}
}
\usage{
data(iowatract20)
}
\description{
Tracts for the state of Iowa
}
\keyword{datasets}
