% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kansastract20.R
\docType{data}
\name{kansastract20}
\alias{kansastract20}
\title{Kansas}
\format{
A data frame with 829 rows and 398 variables:
}
\source{
/url{https://www2.census.gov/programs-surveys/decennial/2020/data/01-Redistricting_File--PL_94-171/}
}
\usage{
data(kansastract20)
}
\description{
Tracts for the state of Kansas
}
\keyword{datasets}
