% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kentuckytract20.R
\docType{data}
\name{kentuckytract20}
\alias{kentuckytract20}
\title{Kentucky}
\format{
A data frame with 1306 rows and 398 variables:
}
\source{
/url{https://www2.census.gov/programs-surveys/decennial/2020/data/01-Redistricting_File--PL_94-171/}
}
\usage{
data(kentuckytract20)
}
\description{
Tracts for the state of Kentucky
}
\keyword{datasets}
