% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/louisianatract20.R
\docType{data}
\name{louisianatract20}
\alias{louisianatract20}
\title{Louisiana}
\format{
A data frame with 1388 rows and 398 variables:
}
\source{
/url{https://www2.census.gov/programs-surveys/decennial/2020/data/01-Redistricting_File--PL_94-171/}
}
\usage{
data(louisianatract20)
}
\description{
Tracts for the state of Louisiana
}
\keyword{datasets}
