% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mainetract20.R
\docType{data}
\name{mainetract20}
\alias{mainetract20}
\title{Maine}
\format{
A data frame with 407 rows and 398 variables:
}
\source{
/url{https://www2.census.gov/programs-surveys/decennial/2020/data/01-Redistricting_File--PL_94-171/}
}
\usage{
data(mainetract20)
}
\description{
Tracts for the state of Maine
}
\keyword{datasets}
