% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marylandtract20.R
\docType{data}
\name{marylandtract20}
\alias{marylandtract20}
\title{Maryland}
\format{
A data frame with 1475 rows and 398 variables:
}
\source{
/url{https://www2.census.gov/programs-surveys/decennial/2020/data/01-Redistricting_File--PL_94-171/}
}
\usage{
data(marylandtract20)
}
\description{
Tracts for the state of Maryland
}
\keyword{datasets}
