% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/massachussetstract20.R
\docType{data}
\name{massachussetstract20}
\alias{massachussetstract20}
\title{Massachussets}
\format{
A data frame with 1620 rows and 398 variables:
}
\source{
/url{https://www2.census.gov/programs-surveys/decennial/2020/data/01-Redistricting_File--PL_94-171/}
}
\usage{
data(massachussetstract20)
}
\description{
Tracts for the state of Massachussets
}
\keyword{datasets}
