% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/michigantract20.R
\docType{data}
\name{michigantract20}
\alias{michigantract20}
\title{Michigan}
\format{
A data frame with 3017 rows and 398 variables:
}
\source{
/url{https://www2.census.gov/programs-surveys/decennial/2020/data/01-Redistricting_File--PL_94-171/}
}
\usage{
data(michigantract20)
}
\description{
Tracts for the state of Michigan
}
\keyword{datasets}
