% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missouritract20.R
\docType{data}
\name{missouritract20}
\alias{missouritract20}
\title{Missouri}
\format{
A data frame with 1654 rows and 398 variables:
}
\source{
/url{https://www2.census.gov/programs-surveys/decennial/2020/data/01-Redistricting_File--PL_94-171/}
}
\usage{
data(missouritract20)
}
\description{
Tracts for the state of Missouri
}
\keyword{datasets}
