% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/montanatract20.R
\docType{data}
\name{montanatract20}
\alias{montanatract20}
\title{Montana}
\format{
A data frame with 319 rows and 398 variables:
}
\source{
/url{https://www2.census.gov/programs-surveys/decennial/2020/data/01-Redistricting_File--PL_94-171/}
}
\usage{
data(alaskatract20)
}
\description{
Tracts for the state of Montana
}
\keyword{datasets}
