% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nebraskatract20.R
\docType{data}
\name{nebraskatract20}
\alias{nebraskatract20}
\title{Nebraska}
\format{
A data frame with 553 rows and 398 variables:
}
\source{
/url{https://www2.census.gov/programs-surveys/decennial/2020/data/01-Redistricting_File--PL_94-171/}
}
\usage{
data(nebraskatract20)
}
\description{
Tracts for the state of Nebraska
}
\keyword{datasets}
