% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nevadatract20.R
\docType{data}
\name{nevadatract20}
\alias{nevadatract20}
\title{Nevada}
\format{
A data frame with 779 rows and 398 variables:
}
\source{
/url{https://www2.census.gov/programs-surveys/decennial/2020/data/01-Redistricting_File--PL_94-171/}
}
\usage{
data(nevadatract20)
}
\description{
Tracts for the state of Nevada
}
\keyword{datasets}
