% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/newhampshiretract20.R
\docType{data}
\name{newhampshiretract20}
\alias{newhampshiretract20}
\title{New Hampshire}
\format{
A data frame with 350 rows and 398 variables:
}
\source{
/url{https://www2.census.gov/programs-surveys/decennial/2020/data/01-Redistricting_File--PL_94-171/}
}
\usage{
data(newhampshiretract20)
}
\description{
Tracts for the state of New Hampshire
}
\keyword{datasets}
