% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/newjerseytract20.R
\docType{data}
\name{newjerseytract20}
\alias{newjerseytract20}
\title{New Jersey}
\format{
A data frame with 2181 rows and 398 variables:
}
\source{
/url{https://www2.census.gov/programs-surveys/decennial/2020/data/01-Redistricting_File--PL_94-171/}
}
\usage{
data(newjerseytract20)
}
\description{
Tracts for the state of New Jersey
}
\keyword{datasets}
