% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/newmexicotract20.R
\docType{data}
\name{newmexicotract20}
\alias{newmexicotract20}
\title{New Mexico}
\format{
A data frame with 612 rows and 398 variables:
}
\source{
/url{https://www2.census.gov/programs-surveys/decennial/2020/data/01-Redistricting_File--PL_94-171/}
}
\usage{
data(newmexicotract20)
}
\description{
Tracts for the state of New Mexico
}
\keyword{datasets}
