% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/newyorktract20.R
\docType{data}
\name{newyorktract20}
\alias{newyorktract20}
\title{New York}
\format{
A data frame with 5411 rows and 398 variables:
}
\source{
/url{https://www2.census.gov/programs-surveys/decennial/2020/data/01-Redistricting_File--PL_94-171/}
}
\usage{
data(newyorktract20)
}
\description{
Tracts for the state of New York
}
\keyword{datasets}
