% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/northcarolinatract20.R
\docType{data}
\name{northcarolinatract20}
\alias{northcarolinatract20}
\title{North Carolina}
\format{
A data frame with 2672 rows and 398 variables:
}
\source{
/url{https://www2.census.gov/programs-surveys/decennial/2020/data/01-Redistricting_File--PL_94-171/}
}
\usage{
data(northcarolinatract20)
}
\description{
Tracts for the state of North Carolina
}
\keyword{datasets}
