% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/puertoricotract20.R
\docType{data}
\name{puertoricotract20}
\alias{puertoricotract20}
\title{Puerto Rico}
\format{
A data frame with 981 rows and 398 variables:
}
\source{
/url{https://www2.census.gov/programs-surveys/decennial/2020/data/01-Redistricting_File--PL_94-171/}
}
\usage{
data(puertoricotract20)
}
\description{
Tracts for the state of Puerto Rico
}
\keyword{datasets}
