% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhodeislandtract20.R
\docType{data}
\name{rhodeislandtract20}
\alias{rhodeislandtract20}
\title{Rhode Island}
\format{
A data frame with 250 rows and 398 variables:
}
\source{
/url{https://www2.census.gov/programs-surveys/decennial/2020/data/01-Redistricting_File--PL_94-171/}
}
\usage{
data(rhodeislandtract20)
}
\description{
Tracts for the state of Rhode Island
}
\keyword{datasets}
