% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/southcarolinatract20.R
\docType{data}
\name{southcarolinatract20}
\alias{southcarolinatract20}
\title{South Carolina}
\format{
A data frame with 1323 rows and 398 variables:
}
\source{
/url{https://www2.census.gov/programs-surveys/decennial/2020/data/01-Redistricting_File--PL_94-171/}
}
\usage{
data(southcarolinatract20)
}
\description{
Tracts for the state of South Carolina
}
\keyword{datasets}
