% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/southdakotatract20.R
\docType{data}
\name{southdakotatract20}
\alias{southdakotatract20}
\title{South Dakota}
\format{
A data frame with 242 rows and 398 variables:
}
\source{
/url{https://www2.census.gov/programs-surveys/decennial/2020/data/01-Redistricting_File--PL_94-171/}
}
\usage{
data(southdakotatract20)
}
\description{
Tracts for the state of South Dakota
}
\keyword{datasets}
