% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tennessetract20.R
\docType{data}
\name{tennessetract20}
\alias{tennessetract20}
\title{Tennesse}
\format{
A data frame with 1701 rows and 398 variables:
}
\source{
/url{https://www2.census.gov/programs-surveys/decennial/2020/data/01-Redistricting_File--PL_94-171/}
}
\usage{
data(tennessetract20)
}
\description{
Tracts for the state of Tennesse
}
\keyword{datasets}
