% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/texastract20.R
\docType{data}
\name{texastract20}
\alias{texastract20}
\title{Texas}
\format{
A data frame with 6896 rows and 398 variables:
}
\source{
/url{https://www2.census.gov/programs-surveys/decennial/2020/data/01-Redistricting_File--PL_94-171/}
}
\usage{
data(texastract20)
}
\description{
Tracts for the state of Texas
}
\keyword{datasets}
