% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vermonttract20.R
\docType{data}
\name{vermonttract20}
\alias{vermonttract20}
\title{Vermont}
\format{
A data frame with 193 rows and 398 variables:
}
\source{
/url{https://www2.census.gov/programs-surveys/decennial/2020/data/01-Redistricting_File--PL_94-171/}
}
\usage{
data(vermonttract20)
}
\description{
Tracts for the state of Vermont
}
\keyword{datasets}
