% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/virginiatract20.R
\docType{data}
\name{virginiatract20}
\alias{virginiatract20}
\title{Virginia}
\format{
A data frame with 2198 rows and 398 variables:
}
\source{
/url{https://www2.census.gov/programs-surveys/decennial/2020/data/01-Redistricting_File--PL_94-171/}
}
\usage{
data(virginiatract20)
}
\description{
Tracts for the state of Virginia
}
\keyword{datasets}
