% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/westvirginiatract20.R
\docType{data}
\name{westvirginiatract20}
\alias{westvirginiatract20}
\title{West Virginia}
\format{
A data frame with 546 rows and 398 variables:
}
\source{
/url{https://www2.census.gov/programs-surveys/decennial/2020/data/01-Redistricting_File--PL_94-171/}
}
\usage{
data(westvirginiatract20)
}
\description{
Tracts for the state of West Virginia
}
\keyword{datasets}
