% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wisconsintract20.R
\docType{data}
\name{wisconsintract20}
\alias{wisconsintract20}
\title{Wisconsin}
\format{
A data frame with 1542 rows and 398 variables:
}
\source{
/url{https://www2.census.gov/programs-surveys/decennial/2020/data/01-Redistricting_File--PL_94-171/}
}
\usage{
data(wisconsintract20)
}
\description{
Tracts for the state of Wisconsin
}
\keyword{datasets}
